Reran roles_dev, substitutions_{dev,paper}: minor changes
The change of language-detection module led to 1188 detected
substitutions instead of 1051, so some values change a little. In
particular, the log-frequency susceptibilities, the #synonyms
susceptibilities, and the #synonyms-#phonemes heatmap, are a bit
changed. Here as in the correlation plots
(2b8dcb85cf56b0800dc8774e324c875304105074), nothing fundamental has
changed (i.e. nothing that impacts our arguments), but it's still worth
noting.