fix/mine: don't drop all cluster substitutions when invalidating one
Previously substitutions were added to the session by backref cascade,
then the session rolled back if a substitution was invalidated. This
lead to dropping all previous substitutions mined in a cluster if a
substitution was invalidated. Disabling cascade_backrefs makes for a
better solution.

Also added tests that failed with the previous code.