fix/mine: use session.rollback() instead of not cascading backrefs
Using cascade_backrefs=False led to a warning when the transaction was
committed with deleted Substitutions in the substitutions_source and
substitutions_destination attributes of quotes. This solution is more
elegant.

Also added a sanity check at the end of the mining process, making sure
we saved exactly all the good subtitutions.