feat/db: compute substitution weights
Substitutions falling on the exact same durl reflect an incertitude on
which exact source was used to create the destination occurrence in
question. So all substitutions with the same durl are pooled together,
and their weight relative to other substitutions in the cluster is given
by `Substitution.durl_weight`.

We also don't want to give advantage to big clusters (in which more
substitutions would take place), so all substitutions are additionally
weighed `1 / number-of-substitutions-in-the-cluster`, so that all
clusters are equal.

Both weights are combined into a substitution's final weight in the
`Substitution.weight` attribute.