feat/mine: improve substitution filtering and test it
We don't want any substitutions that are in fact the deletion (or
re-composition) of a word, caught because we allow substrings. I.e. we
don't want 'I am in high school' -> 'I am in school' to count as a
substitution, not do we want 'policy makers are fools' -> 'policymakers
are fools' to count. We filter these out.