feat/susceptibility: improve computation coherence
Several important changes:
* introduce H0s for all susceptibility measures, and compare to those
* fix the computation for in-sentence quantiles

This is a bit of a checkpoint, because in the next commit I intend to
change the way susceptibilities are computed: instead of computing a
susceptibility value per feature value, and averaging in bins, I'll do
the same as for in-sentence quantiles, that is count the number of times
a feature value appears (scaled to sentence length) vs. the number of
times it is the target of a substitution, creating a proper H0 this way.
This is because, with this computation, H0 is really 1, whereas in the
previous computation (the one in this commit), H0 is not exactly 1
necessarily.