feat/susceptibility: make computation with H0s much more coherent
As explained in the previous commit, instead of computing a
susceptibility value for each feature value, and averaging over that, we
now compute a susceptibility value for each bin. This lets us have
coherent H0 values, because we know the sum of susceptibilities under H0
should the same as in the experiment (see explanations in the notebook
for more details).

Finally, we also weigh substitutions according to the number of other
substitutions they appear with, *where the source word is coded*
(otherwise we're dividing by more than necessary, counting substitutions
that aren't coded by our current feature).